[#26] refactor: renaming custom error DTO
