+++ src/modules/models/customError/types/ICreateCustomErrorDTO.ts
